% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_chart.R
\name{add_chart}
\alias{add_chart}
\title{Add chart config (adapted from golem::add_module())}
\usage{
add_chart(
  path,
  name = "newplot",
  label = "New Static Plot",
  type = "plot",
  domain = "labs",
  package = NULL,
  workflow = list(),
  open = TRUE,
  ...
)
}
\arguments{
\item{path}{Path to store the chart configuration files, expecting a config root folder

+-- config
|  +-- aeExplorer.yaml
|  +-- newChart.yaml

|      +-- aeExplorer_init.R
|      +-- newChart_main.R}

\item{name}{The name of the chart (also name of the yaml file)}

\item{label}{Label of chart}

\item{type}{Type of chart: \code{plot}, \code{module}, or \code{htmlwidget}. Default is \code{plot} (static)}

\item{domain}{associated data domain, for example \code{aes}, \code{labs}, or \code{multiple}}

\item{package}{optional, R package that this chart is associated with.}

\item{workflow}{list of workflow functions appropriate for chart YAML}

\item{open}{boolean}

\item{...}{additional parameters for chart_template}
}
\description{
This function creates a module inside the local \verb{config/} folder to define a new chart
\itemize{
\item dump yaml
\item create R scripts
}
}
\seealso{
\code{\link[=chart_template]{chart_template()}}
}
