% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addChart.R
\name{addChart}
\alias{addChart}
\title{Adds a new chart to the safetyGraphics shiny app}
\usage{
addChart(chart, label = "", description = "", repo_url = "",
  settings_url = "", main = "character", type = "htmlwidget",
  maxWidth = 1000, requiredSettings = c(""),
  settingsLocation = getwd(), overwrite = FALSE)
}
\arguments{
\item{chart}{Name of the chart - one word, all lower case}

\item{label}{Nicely formatted name of the chart}

\item{description}{Description of the chart}

\item{repo_url}{Homepage for chart's code repository (if any)}

\item{settings_url}{Homepage for chart's settings documentation}

\item{main}{Name of the main function used to initialize the app. If the type is htmlwidgets, the js function must accept "location" and "settings" parameters (in that order) and have an .init() method, expecting a json data array. Otherwise, the r function should accept named data and settings parameters, andshould be loaded in the user's namespace.}

\item{type}{type of chart (e.g. 'htmlwidget')}

\item{maxWidth}{max width for the widget in pixels}

\item{requiredSettings}{array of text_key values (matching those used in settingsMetadata) for the required settings for this chart}

\item{settingsLocation}{folder location of user-defined settings metadata}

\item{overwrite}{overwrite any existing chart metadata? default = false}
}
\description{
This function updates settings objects to adds a new chart to the safetyGraphics shiny app
}
\details{
This function makes it easy for users to adds a new chart to the safetyGraphics shiny app, by making updates to the underlying metadata used by the package. Specifically, the function adds a row to chartsMetadata.rda describing the chart and adds a column to settingsMetadata.rda specifying which settings are used with the chart. If new settings are needed for the chart, the user should call addSetting() for each new setting required.
}
