% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlbc.R
\docType{data}
\name{adlbc}
\alias{adlbc}
\title{ADLBC Dataset}
\format{A data frame with 10288 rows and 46 variables.}
\source{
Rho, Inc.
}
\usage{
adlbc
}
\description{
A dataset containing things that Becca and Jeremy will better explain
}
\keyword{datasets}
