% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\docType{data}
\name{aes}
\alias{aes}
\title{Adverse Event sample data}
\format{
A data frame with 10288 rows and 46 variables.
\describe{
\item{STUDYID}{Study Identifier}
\item{SUBJID}{Subject Identifier for the Study}
\item{USUBJID}{Unique Subject Identifier}
}
}
\source{
\url{https://github.com/RhoInc/data-library}
}
\usage{
aes
}
\description{
A dataset containing anonymized AE data from a clinical trial in the CDISC ADaM format. The structure is 1 record per measure per adverse event. See a full description of the ADaM data standard \href{https://www.cdisc.org/standards/foundational/adam/adam-implementation-guide-v11}{here}.
}
\keyword{datasets}
