% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_init_addin.R
\name{app_init_addin}
\alias{app_init_addin}
\title{RStudio Add-in for constructing lean ADLB and ADAE data}
\usage{
app_init_addin()
}
\description{
RStudio Add-in for constructing lean ADLB and ADAE data
}
