% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Server for the default safetyGraphics shiny app}
\usage{
app_server(meta, mapping, domainData, charts, filterDomain)
}
\arguments{
\item{meta}{data frame containing the metadata for use in the app. See the preloaded file (\code{?safetyGraphics::meta}) for more data specifications and details. Defaults to \code{safetyGraphics::meta}.}

\item{mapping}{current mapping}

\item{domainData}{named list of data.frames to be loaded in to the app.}

\item{charts}{list of charts to include in the app}

\item{filterDomain}{domain used for the data/filter tab. Demographics ("\code{dm}") is used by default. Using a domain that is not one record per participant is not recommended.}
}
\description{
This function returns a server function suitable for use in shiny::runApp()
}
