% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_chart.R
\name{chart_template}
\alias{chart_template}
\title{Chart Template Function}
\usage{
chart_template(name, path, type, ...)
}
\arguments{
\item{name}{The name of the chart (also name of the yaml file)}

\item{path}{The path to the R script where the module will be written.
Note that this path will not be set by the user but internally by
\code{add_chart()}.}

\item{type}{Type of chart: \code{plot}, \code{module}, or \code{htmlwidget}. Default is \code{plot} (static)}

\item{...}{Arguments to be passed to the template, via \code{add_chart()}}
}
\value{
Used for side effect
}
\description{
Chart Template Function
}
\seealso{
\code{\link[=add_chart]{add_chart()}}
}
