% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartsMetadata.R
\docType{data}
\name{chartsMetadata}
\alias{chartsMetadata}
\title{Charts Metadata}
\format{A data frame with 25 rows and 3 columns
\describe{
   \item{chart}{Name of the chart - one word, all lower case}
   \item{label}{Nicely formatted name of the chart}
   \item{description}{Description of the chart}
   \item{repo_url}{Homepage for chart's code repository (if any)}
   \item{settings_url}{Homepage for chart's settings documentation}
   \item{type}{type of chart (e.g. 'htmlwidget')}
}}
\source{
Created for this package
}
\usage{
chartsMetadata
}
\description{
Metadata about the charts available in the shiny app
}
\keyword{datasets}
