% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chartsNav.R
\name{chartsNav}
\alias{chartsNav}
\title{Add a navbar tab that initializes the Chart Module UI}
\usage{
chartsNav(name, label, type, package)
}
\arguments{
\item{chart}{chart metadata}
}
\description{
Add a navbar tab that initializes the Chart Module UI
}
