% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chartsNav.R
\name{chartsNavUI}
\alias{chartsNavUI}
\title{Adds a navbar tab that initializes the Chart Module UI}
\usage{
chartsNavUI(id, chart)
}
\arguments{
\item{id}{module id}

\item{chart}{chart metadata}
}
\description{
Adds a navbar tab that initializes the Chart Module UI
}
