% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chartsRenderModule.R
\name{chartsRenderModuleUI}
\alias{chartsRenderModuleUI}
\title{Charts Module - render module chart UI}
\usage{
chartsRenderModuleUI(id, customModUI)
}
\description{
Charts Module - sub module for rendering a static chart
}
