% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chartsRenderStatic.R
\name{chartsRenderStatic}
\alias{chartsRenderStatic}
\title{Charts Module - render static chart server}
\usage{
chartsRenderStatic(input, output, session, chartFunction, params, type)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{chartFunction}{function to generate the chart.}

\item{params}{parameters to be passed to the widget (Reactive)}
}
\description{
server for the display of the loaded data
}
