% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chartsRenderStatic.R
\name{chartsRenderStaticUI}
\alias{chartsRenderStaticUI}
\title{Charts Module - render static chart UI}
\usage{
chartsRenderStaticUI(id)
}
\description{
Charts Module - sub module for rendering a static chart
}
