% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chartsRenderWidget.R
\name{chartsRenderWidget}
\alias{chartsRenderWidget}
\title{Charts Module - render static chart server}
\usage{
chartsRenderWidget(
  input,
  output,
  session,
  widgetName,
  widgetPackage = "safetyGraphics",
  initFunction,
  data,
  mapping
)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{widgetName}{str name of the widget}

\item{widgetPackage}{str package containing the widget}

\item{initFunction}{function called before the chart is generated. Should return a list of parameters that will be provided to chartFunction. returns list(data=data, settings=settings) by default.}

\item{data}{named list of data sets [reactive]}

\item{mapping}{data mapping [reactive]}
}
\description{
server for the display of the loaded data
}
