% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chartsRenderWidget.R
\name{chartsRenderWidgetUI}
\alias{chartsRenderWidgetUI}
\title{Charts Module - render static chart UI}
\usage{
chartsRenderWidgetUI(id, widgetName, widgetPackage = "safetyGraphics")
}
\description{
Charts Module - sub module for rendering a static chart
}
