% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chartsTab.R
\name{chartsTabUI}
\alias{chartsTabUI}
\title{Charts Tab}
\usage{
chartsTabUI(id, chart)
}
\arguments{
\item{id}{module id}

\item{chart}{list containing chart specifications}
}
\description{
Charting module
}
