% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chartsTab.R
\name{chartTab}
\alias{chartTab}
\title{home tab - server}
\usage{
chartTab(input, output, session, chart, data, mapping)
}
\arguments{
\item{input}{Input objects from module namespace}

\item{output}{Output objects from module namespace}

\item{session}{An environment that can be used to access information and functionality relating to the session}

\item{data}{list of data frames  [REACTIVE]}

\item{mapping}{data frame containing mapping arguments for chart [REACTIVE]}

\item{chartID}{string containing the chart ID.}

\item{chartMeta}{data frame with the following columns. \enumerate{
\item `chartID` chartID
\item `label` text label
\item `type` type must be  'htmlwidget', 'module', 'static' or 'plotly'. Other optional parameters are: 
\item `UI` a function that accepts parameters called `data` and `mapping` and returns a chart object of the expected type. Defaults to `name()`
\item `server` a function that calls a shiny module server defaults to `nameServer()`. Ignored unless type='module'.
\item `onInit` a function that that accepts parameters called `data` and `mapping` and returns a list of custom parameters to be passed to the UI and server functions. returns list(data=data, mapping=mapping) with no changes by default. 
} If chartID isn't found in chartMeta, the system will attempt to draw a static plot with the provided data and mapping.}
}
\description{
server for the display of the home tab
}
