% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_columnMapping.R
\name{columnMappingInput}
\alias{columnMappingInput}
\title{ColumnMappingUI and columnMapping}
\usage{
columnMappingInput(id, choices, selected)
}
\description{
Module that facilitates the mapping of a single column with a simple select UI
}
\section{Output}{

\describe{
\item{\code{columnMapping}}{list describing the column mapping}
}
}

