% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_safetyGraphics_app.R
\name{create_new_safetyGraphics_app}
\alias{create_new_safetyGraphics_app}
\title{start new project with an instance of safetyGraphicsApp()}
\usage{
create_new_safetyGraphics_app(
  path,
  init_default_configs,
  open = TRUE,
  gui = FALSE
)
}
\arguments{
\item{path}{location for new safetyGraphicsApp}

\item{init_default_configs}{copy over \code{safetyCharts} default configs?}

\item{open}{open new rstudio project?}

\item{gui}{show gui? creates rstudio project if false}
}
\value{
Used for side effect
}
\description{
start new project with an instance of safetyGraphicsApp()
}
