% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createChart.R
\name{createChart}
\alias{createChart}
\title{Create a custom static or plotly R graphic}
\usage{
createChart(type, rSettings)
}
\arguments{
\item{type}{Type of chart. Options are \code{"static"} or \code{plotly}.}

\item{rSettings}{List containing all arguments needed for chart creation: 
\code{rSettings = list(
data = data,
debug_js=debug_js,
chartFunction = chartFunction,
settings = jsonlite::toJSON(
  settings,
  auto_unbox = TRUE,
  null = "null"
))}}
}
\description{
Create a custom static or plotly R graphic
}
\keyword{internal}
