% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\docType{data}
\name{dm}
\alias{dm}
\title{Demographics sample data}
\format{
A data frame with 306 rows and 25 variables.
}
\source{
\url{https://github.com/RhoInc/data-library}
}
\usage{
dm
}
\description{
A dataset containing anonymized demographcis data from a clinical trial in the CDISC ADaM format. The structure is 1 record per measure per adverse event. See a full description of the ADaM data standard \href{https://www.cdisc.org/standards/foundational/adam/adam-implementation-guide-v11}{here}.
}
\keyword{datasets}
