% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edish.R
\name{edish}
\alias{edish}
\title{Create an edish widget}
\usage{
edish(data, debug_js = FALSE, settings = NULL)
}
\arguments{
\item{data}{A data frame containing the labs data. Data must be structured as one record per study participant per time point per lab measure.}

\item{debug_js}{print settings in javascript before rendering chart. Default: \code{FALSE}.}

\item{settings}{List of settings arguments to be converted to JSON using: \preformatted{
jsonlite::toJSON(
  settings,
  auto_unbox = TRUE,
  dataframe = "rows",
  null = "null"
)}

Default: \code{NULL}.}
}
\description{
This function creates an \href{https://github.com/SafetyGraphics/safety-eDISH/}{interactive graphic} for the Evaluation of Drug-Induced Serious Hepatotoxicity (edish). See the \href{https://github.com/SafetyGraphics/safety-eDISH/wiki/Configuration}{chart docuemtnation} for details regarding the settings object.
}
\examples{
\dontrun{

## Create edish figure using a premade settings list
group_cols_list <- list(
  list(value_col = "TRTP", label = "Treatment"),
  list(value_col = "SEX", label = "Sex"),
  list(value_col = "AGEGR1", label = "Age group")
)


filters_list <- list(
  list(value_col = "TRTA", label = "Treatment"),
  list(value_col = "SEX", label = "Sex"),
  list(value_col = "RACE", label = "RACE"),
  list(value_col = "AGEGR1", label = "Age group")
)

settingsl <- list(id_col = "USUBJID",
      value_col = "AVAL",
      measure_col = "PARAM",
      visit_col = "VISIT",
      visitn_col = "VISITNUM",
      studyday_col = "ADY",
      normal_col_low = "A1LO",
      normal_col_high = "A1HI",
      group_cols = group_cols_list,
      filters = filters_list,
      measure_values = list(ALT = "Alanine Aminotransferase (U/L)",
                            AST = "Aspartate Aminotransferase (U/L)",
                            TB = "Bilirubin (umol/L)",
                            ALP = "Alkaline Phosphatase (U/L)"))
edish(data=adlbc, settings = settingsl)

}

}
