% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filterTabChecks.R
\name{filterTabChecks}
\alias{filterTabChecks}
\title{filter module checks}
\usage{
filterTabChecks(domainData, filterDomain, current_mapping)
}
\arguments{
\item{domainData}{list of data files for each domain}

\item{filterDomain}{domain to use for filtering (typically "dm")}

\item{current_mapping}{current data mapping (REACTIVE)}

\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\value{
reactive that returns a boolean indicating whether the checks passed and filtering can be initiailized
}
\description{
function that checks whether the current data and settings are appropriate for the filter tab
}
