% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filterTab.R
\name{filterTabUI}
\alias{filterTabUI}
\title{UI for the filter module in datamods::filter_data_ui}
\usage{
filterTabUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
UI for the filter module in datamods::filter_data_ui
}
