% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMappingList.R
\name{generateMappingList}
\alias{generateMappingList}
\title{Convert mapping data.frame to a list}
\usage{
generateMappingList(settingsDF, domain = NULL, pull = FALSE)
}
\arguments{
\item{settingsDF}{data frame containing current mapping}

\item{domain}{mapping domain to return (returns all domains as a named list by default)}

\item{pull}{call pull() the value for each parameter - needed for testing only. default: FALSE}
}
\description{
Convert mapping data.frame to a list
}
