% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateShell.R
\name{generateShell}
\alias{generateShell}
\title{Generate a default settings shell based on settings metadata}
\usage{
generateShell(charts = "eDish")
}
\arguments{
\item{charts}{The chart or chart(s) for which shells should be generated ("eDish" only for now) . Default: \code{"eDish"}.}
}
\value{
A list containing the appropriate settings for the selected chart
}
\description{
This function returns a default settings objectbased on the chart(s) specified.
}
\details{
The function is designed to work with valid safetyGraphics charts.
}
\examples{

generateShell(chart = "eDish") 
 
}
\keyword{internal}
