% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChartFunctions.R
\name{getChartFunctions}
\alias{getChartFunctions}
\title{Get Chart Functions}
\usage{
getChartFunctions(chartsList, chartSettingsPaths)
}
\arguments{
\item{chartsList}{List of charts and thier parameters}

\item{chartSettingsPath}{path for custom files.}
}
\value{
an updated chartsList object with initFunction and chartFunction parameters added as appropriate
}
\description{
Function to get functions associated with charts
}
