% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChartSettings.R
\name{getChartSettings}
\alias{getChartSettings}
\title{Get chart-specific settings}
\usage{
getChartSettings(settings, chart)
}
\arguments{
\item{settings}{Settings list containing settings for all selected charts.}

\item{chart}{The chart for which settings should be returned.}
}
\value{
Chart-specific settings
}
\description{
Subset master settings list to chart-specific settings list only
}
\examples{
settings <- safetyGraphics::generateSettings(standard="ADaM")
safetyGraphics:::getChartSettings(settings = settings, chart = "edish")

}
\keyword{internal}
