% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChartStatus.R
\name{getChartStatus}
\alias{getChartStatus}
\title{Check the status of a chart based on the current mapping}
\usage{
getChartStatus(chart, mapping)
}
\arguments{
\item{chart}{chart object}

\item{mapping}{the current mapping data.frame}
}
\value{
a list with \code{status}, \code{domains} and \code{columns} properties
}
\description{
Checks a chart's status when column-level chart specifications are provided in chart$dataSpec. Note that safetyGraphicsApp() does not allow a \code{mapping} value that is not found in \code{domainData}, so this function only needs to check that an expected parameter exists in \code{mapping} (not that the specified column is found in the loaded data).
}
\details{
Returns a list, with:
\itemize{
\item \code{status} (TRUE, FALSE)
\item \code{domains} a list specifying wheter all columns are specified in each domain
\item \code{columns}  a list that matches the structure of chart$dataSpec and indicates which variables are available.
}
}
\examples{
sample_chart <- list(
    domains=c("aes","dm"),
    dataSpec=list(
        aes=c("id_col","custom_col"),
        dm=c("id_col","test_col")
    )
)

sample_mapping <- data.frame(
    domain=c("aes","aes","dm","dm"),
    text_key=c("id_col","custom_col","id_col","test_col"),
    current=c("myID","AEcol","myID","dmCol")
)

check <- safetyGraphics:::getChartStatus(chart=sample_chart, mapping=sample_mapping) 
# check$status=TRUE

}
\keyword{internal}
