% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{Get combined safetyGraphics metadata environment}
\usage{
getMetadata(meta = safetyGraphics::metadata, path = NULL, domain = NULL)
}
\arguments{
\item{meta}{Starting metadata object. Defaults to safetyGraphics::metadata}

\item{path}{Optional location a custom metadata file. Data loaded from path will overwrite data loaded via meta.}

\item{domain}{Optional string (or vector of strings) used to filter metadata to one or more data domains. All domains returned by default.}
}
\value{
list with metdata following the format specified in safetyGraphics:::metadata
}
\description{
Create a list containing all metadata (settings, charts and standards) needed for creating charts in safetyGraphics
}
\examples{
all <- getMetadata() #returns safetygraphics:::metadata
labs <- getMetadata(domian="labs") #returns lab domain only
custom <- file.path(getwd(), 'metadata.rds')  #attempt to load `metadata.rds` saved in the working directory

}
