% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRequiredColumns.R
\name{getRequiredColumns}
\alias{getRequiredColumns}
\title{Get a list of required columns}
\usage{
getRequiredColumns(standard, chart = "eDish")
}
\arguments{
\item{standard}{The data standard for which to fetch required columns Valid options are "SDTM", "AdAM".}

\item{chart}{The chart for which standards should be generated ("eDish" only for now) . Default: \code{"eDish"}.}
}
\value{
A character vector of required data columns
}
\description{
Get a list of required columns for a chart in a given data standard
}
\examples{
safetyGraphics:::getRequiredColumns(standard="ADAM")
safetyGraphics:::getRequiredColumns(standard="SDTM")

}
\keyword{internal}
