% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_homeTab.R
\name{homeTab}
\alias{homeTab}
\title{home tab - server}
\usage{
homeTab(input, output, session)
}
\arguments{
\item{input}{mod input}

\item{output}{mod output}

\item{session}{mod session}
}
\description{
server for the display of the home tab
}
