% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loadCharts.R
\name{loadCharts}
\alias{loadCharts}
\title{Server for the chart loading module used in safetyGraphicsInit()}
\usage{
loadCharts(input, output, session, charts = makeChartConfig())
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{charts}{list containing chart specifications like those returned by \link{makeChartConfig}.}
}
\description{
Server for the chart loading module used in safetyGraphicsInit()
}
