% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loadCharts.R
\name{loadChartsUI}
\alias{loadChartsUI}
\title{UI for the chart loading module used in safetyGraphicsInit()}
\usage{
loadChartsUI(id, charts = makeChartConfig())
}
\arguments{
\item{id}{module id}

\item{charts}{list containing chart specifications like those returned by \link{makeChartConfig}.}
}
\description{
UI for the chart loading module used in safetyGraphicsInit()
}
