% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loadData.R
\name{loadData}
\alias{loadData}
\title{Server for the data loading module used in safetyGraphicsInit()}
\usage{
loadData(input, output, session, domain)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{domain}{data domain to be loaded}
}
\description{
Server for the data loading module used in safetyGraphicsInit()
}
