% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loadDomains.R
\name{loadDomains}
\alias{loadDomains}
\title{loadDataServer}
\usage{
loadDomains(input, output, session, domains)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{domains}{List of data domains to be loaded {reactive}}
}
\description{
UI that facilitates the mapping of a column data domain
}
