% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loadDomains.R
\name{loadDomainsUI}
\alias{loadDomainsUI}
\title{loadDataUI}
\usage{
loadDomainsUI(id)
}
\arguments{
\item{id}{module id}

\item{domains}{character vector with initial domains to be loaded}
}
\description{
UI that facilitates the mapping of a column data domain
}
