% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeChartConfigFunctions.R
\name{makeChartConfigFunctions}
\alias{makeChartConfigFunctions}
\title{Make Chart Config Functions}
\usage{
makeChartConfigFunctions(chart)
}
\arguments{
\item{chart}{chart object like the one generated by makeChartConfig().}
}
\value{
returns the chart object with a new functions object added.
}
\description{
Binds needed functions to a chart object based on chart type.
}
