% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeChartExport.R
\name{makeChartExport}
\alias{makeChartExport}
\title{Make Chart Export}
\usage{
makeChartExport(chart, mapping)
}
\arguments{
\item{chart}{chart object like the one generated by makeChartConfig().}

\item{mapping}{mapping object like the one generated by makeMapping().}
}
\value{
returns a character vector that can be saved as R code.
}
\description{
Creates R code that allows chart to be exported
}
