% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeChartParams.R
\name{makeChartParams}
\alias{makeChartParams}
\title{Make Chart Parameters}
\usage{
makeChartParams(data, chart, mapping)
}
\arguments{
\item{chart}{list containing chart specifications}

\item{mapping}{data frame with current mapping}

\item{dataDomains}{list of domain-level data}
}
\description{
Updates raw data and mapping for use with a specific chart
}
