% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMapping.R
\name{makeMapping}
\alias{makeMapping}
\title{Create data mapping based on data standards and user input}
\usage{
makeMapping(domainData, meta, autoMapping, customMapping)
}
\arguments{
\item{domainData}{named list of data.frames to be loaded in to the app. Sample AdAM data from the safetyData package used by default}

\item{meta}{data frame containing the metadata for use in the app. See the preloaded file (\code{?safetyGraphics::meta}) for more data specifications and details. Defaults to \code{safetyGraphics::meta}.}

\item{autoMapping}{boolean indicating whether the app should attempt to automatically detect data standards and generate mappings for the data provided. Values specified in the \code{mapping} parameter overwrite automatically generated mappings when both are found. Defaults to true.#'}

\item{mapping}{list specifying the initial mapping values for each data mapping for each domain (e.g. list(aes= list(id_col='USUBJID', seq_col='AESEQ')).}
}
\value{
List of elements for used to initialize the shiny app with the following parameters
\itemize{
\item{"meta"}{ List of configuration metadata }
\item{"charts"}{ List of charts }
\item{"domainData"}{ List of domain level data sets }
\item{"mapping"}{ Initial Data Mapping  }
\item{"standards"}{ List of domain level data standards }
}
}
\description{
Create data mapping based on data standards and user input
}
