% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMeta.R
\name{makeMeta}
\alias{makeMeta}
\title{Create a metadata object based on the selected charts}
\usage{
makeMeta(charts, package = "safetyCharts")
}
\arguments{
\item{charts}{list of charts for which metadata is needed}

\item{package}{package containing needed metadata}
}
\value{
tibble of metadata with the following columns:
\describe{
\item{domain}{Data domain}
\item{text_key}{Text key indicating the setting name. \code{'--'} delimiter indicates a field level data mapping}
\item{col_key}{Key for the column mapping}
\item{field_key}{Key for the field mapping (if any)}
\item{type}{type of mapping - "field" or "column"}
\item{label}{Label}
\item{description}{Description}
\item{multiple}{Mapping supports multiple columns/fields }
\item{standard_adam}{Default values for the ADaM data standard}
\item{standard_sdtm}{Default values for the SDTM data standard}
}
}
\description{
Create a metadata object based on the selected charts
}
