% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mappingColumn.R
\name{mappingColumn}
\alias{mappingColumn}
\title{Server that facilitates the mapping of a column data (and any associated fields)}
\usage{
mappingColumn(input, output, session, meta, data)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{meta}{metadata data frame for the object}

\item{data}{current data file for the domain}
}
\value{
A reactive data.frame providing the current value for text_key associated with the selected column
}
\description{
Server that facilitates the mapping of a column data (and any associated fields)
}
