% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mappingColumn.R
\name{mappingColumnUI}
\alias{mappingColumnUI}
\title{UI that facilitates the mapping of a column data (and any associated fields)}
\usage{
mappingColumnUI(id, meta, data, mapping = NULL)
}
\arguments{
\item{id}{module id}

\item{meta}{metadata for the column (and related fields)}

\item{data}{current data file for the domain}

\item{mapping}{current data mapping for the column (and related fields)}
}
\description{
UI that facilitates the mapping of a column data (and any associated fields)
}
