% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mappingDomain.R
\name{mappingDomain}
\alias{mappingDomain}
\title{Server that facilitates the mapping of a full data domain}
\usage{
mappingDomain(input, output, session, meta, data)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{meta}{metadata for the domain}

\item{data}{clinical data for the domain}
}
\value{
A reactive data frame containing the mapping for the domain
}
\description{
Server that facilitates the mapping of a full data domain
}
