% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mappingDomain.R
\name{mappingDomainUI}
\alias{mappingDomainUI}
\title{UI that facilitates the mapping of a full data domain}
\usage{
mappingDomainUI(id, meta, data, mapping = NULL)
}
\arguments{
\item{id}{module id}

\item{meta}{metadata for the domain}

\item{data}{data file for the domain}

\item{mapping}{current data mapping}
}
\description{
UI that facilitates the mapping of a full data domain
}
