% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mappingSelect.R
\name{mappingSelect}
\alias{mappingSelect}
\title{Server that facilitates the mapping of a single data element (column or field) with a simple select UI}
\usage{
mappingSelect(input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\value{
A reactive containing the selected column
}
\description{
Server that facilitates the mapping of a single data element (column or field) with a simple select UI
}
