% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mappingTab.R
\name{mappingTabUI}
\alias{mappingTabUI}
\title{mappingTabUI}
\usage{
mappingTabUI(id, meta, domainData, mappings = NULL, standards = NULL)
}
\arguments{
\item{meta}{metadata for all domains}

\item{domainData}{list of data files for each domain}

\item{mappings}{data frame containing stacked mappings for all domains}
}
\description{
UI that facilitates the mapping of all data domain
}
