% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Metadata}
\format{
A list containing 3 named data frames (`settings`, `charts` and `standards`) containing metadata needed for renderering custom charts using safetyGraphics.
 
\describe{
    \item{settings$chart_hepexplorer}{Flag indicating if the settings apply to the Hepatic Explorer Chart}
   \item{settings$chart_paneledoutlierexplorer}{Flag indicating if the settings apply to the Paneled Safety Outlier Explorer Chart}
   \item{settings$chart_safetyhistogram}{Flag indicating if the settings apply to the Safety Histogram Chart}
   \item{settings$chart_safetyoutlierexplorer}{Flag indicating if the settings apply to the Safety Outlier Explorer Chart}
   \item{settings$chart_safetyresultsovertime}{Flag indicating if the settings apply to the Safety Results Over Time Chart}
   \item{settings$chart_safetyshiftplot}{Flag indicating if the settings apply to the Safety Shift Plot Chart} 
   \item{settings$chart_safetydeltadelta}{Flag indicating if the settings apply to the Safety Delta-Delta Chart} 
   \item{settings$domain}{Data domain for the setting}
   \item{settings$text_key}{Text key indicating the setting name. \code{'--'} delimiter indicates a nested setting}
   \item{settings$text_key}{Text key indicating the setting name. \code{'--'} delimiter indicates a nested setting}
   \item{settings$label}{Label}
   \item{settings$description}{Description}
   \item{settings$setting_type}{Expected type for setting value. Should be "character", "vector", "numeric" or "logical"}
   \item{settings$setting_required}{Flag indicating if the setting is required}
   \item{settings$column_mapping}{Flag indicating if the setting corresponds to a column in the associated data}
   \item{settings$column_type}{Expected type for the data column values. Should be "character","logical" or "numeric"}
   \item{settings$field_mapping}{Flag indicating whether the setting corresponds to a field-level mapping in the data}
   \item{settings$field_column_key}{Key for the column that provides options for the field-level mapping in the data}
   \item{settings$setting_cat}{Setting category (data, measure, appearance)}
   \item{settings$default}{Default value for non-data settings}
   \item{chart$domain}{Data domain for the chart}
   \item{charts$chart}{Name of the chart - one word, all lower case}
   \item{charts$label}{Nicely formatted name of the chart}
   \item{charts$description}{Description of the chart}
   \item{charts$repo_url}{Homepage for chart's code repository (if any)}
   \item{charts$settings_url}{Homepage for chart's settings documentation}
   \item{charts$main}{Name of the main function used to initialize the app. The function must accept "location" and "settings" parameters (in that order) and have an .init() method, expecting a json data array.}
   \item{charts$type}{type of chart (e.g. 'htmlwidget')}
   \item{charts$maxWidth}{max width for the widget in pixels}
   \item{standards$text_key}{Text key indicating the setting name. \code{'--'} delimiter indicates a nested setting}
   \item{standards$adam}{Settings values for the ADaM standard}
   \item{standards$sdtm}{Settings values for the SDTM standard}
}
}
\source{
Created for this package
}
\usage{
metadata
}
\description{
Metadata about the charts, data standards and settings available in the shiny app. The settings table (metadata$settings)  contains information about the settings used to configure safetyGraphics charts and has one record per unique setting. The standards table (metadata$standards) contains information about the data standards used to configure safetyGraphics charts and has one record per unique setting. Columns contain default setting values for clinical data standards, like the CDISC "adam" and "sdtm" standards. The charts table (metadata$charts) has information about the charts available in the package and has one unique record per chart type.
}
\keyword{datasets}
