% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mappingMeta.R
\name{metaMappingUI}
\alias{metaMappingUI}
\title{Metadata Mapping UI}
\usage{
metaMappingUI(id)
}
\description{
UI that displays the data mapping metadata.
}
\section{Output}{

\describe{
\item{\code{mappingMeta}}{Reactive containing the metadata to be used in the app.}
}
}

