% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareChart.R
\name{prepareChart}
\alias{prepareChart}
\title{Prepare a chart object for safetyGraphcis}
\usage{
prepareChart(chart)
}
\arguments{
\item{chart}{chart object like the one generated by makeChartConfig().}
}
\value{
returns the chart object with a new functions object added.
}
\description{
Sets default values and binds needed functions to a chart object based on chart type.
}
