% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_profileTab.R
\name{profileTab}
\alias{profileTab}
\title{Server for the filter module in safetyProfile::profile_server}
\usage{
profileTab(input, output, session, params)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{params}{reactive containing mapping and data}

\item{current_id}{reactive containing currently selected participant}
}
\value{
current_id
}
\description{
Server for the filter module in safetyProfile::profile_server
}
