% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_profileTab.R
\name{profileTabUI}
\alias{profileTabUI}
\title{UI for the profile module in safetyProfile::profile_ui}
\usage{
profileTabUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
UI for the profile module in safetyProfile::profile_ui
}
