% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeCharts.R
\name{removeCharts}
\alias{removeCharts}
\title{Remove a chart from the safetyGraphics shiny app}
\usage{
removeCharts(domain, charts, metadataLocation = getwd())
}
\arguments{
\item{domain}{domain of the chart}

\item{charts}{Name of the chart(s) to remove - one word, all lower case}

\item{metadataLocation}{Path where the custom settings will be loaded/saved.}
}
\description{
This function updates settings objects to remove chart from the safetyGraphics shiny app
}
\details{
This function makes it easy for remove a chart from the safetyGraphics shiny app by making updates to the underlying metadata used by the package.
}
