% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSettings.R
\name{removeSettings}
\alias{removeSettings}
\title{Remove a setting from the safetyGraphics shiny app}
\usage{
removeSettings(domain, text_keys, settingsLocation = getwd())
}
\arguments{
\item{domain}{domain of the chart}

\item{text_keys}{Text keys indicating the setting names to be removed.}

\item{metadataLocation}{Path where the custom settings will be loaded/saved.}
}
\description{
This function updates settings objects to remove a setting parameter from the safetyGraphics shiny app
}
\details{
This function makes it easy for remove a setting from the safetyGraphics shiny app by making updates to the underlying metadata used by the package.
}
