% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_reportsTab.R
\name{reportsTab}
\alias{reportsTab}
\title{Reports tab - server}
\usage{
reportsTab(input, output, session, charts, data, mapping)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{charts}{list containing safetyGraphics chart objects. see custom chart vignette for details.}

\item{data}{named list of current data sets \link{reactive}.}

\item{mapping}{tibble capturing the current data mappings \link{reactive}.}
}
\description{
server for the chart export module
}
