% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyGraphics.R
\docType{package}
\name{SafetyGraphics}
\alias{SafetyGraphics}
\title{\code{safetyGraphics} package}
\description{
Create Interactive Graphics Related to Clinical Trial Safety
}
\details{
Learn more on
\href{https://CRAN.R-project.org/package=safetyGraphics}{CRAN}
or \href{https://github.com/ASA-DIA-InteractiveSafetyGraphics/safetyGraphics}{GitHub}
}
\keyword{internal}
