% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_safetyGraphicsUI.R
\name{safetyGraphicsUI}
\alias{safetyGraphicsUI}
\title{UI for the default safetyGraphics shiny app}
\usage{
safetyGraphicsUI(id, meta, domainData, mapping, standards)
}
\arguments{
\item{id}{module ID}

\item{meta}{data frame containing the metadata for use in the app. See the preloaded file (\code{?safetyGraphics::meta}) for more data specifications and details. Defaults to \code{safetyGraphics::meta}.}

\item{domainData}{named list of data.frames to be loaded in to the app.}

\item{mapping}{data.frame specifying the initial values for each data mapping. If no mapping is provided, the app will attempt to generate one via \code{detectStandard()}}

\item{standards}{a list of information regarding data standards. Each list item should use the format returned by safetyGraphics::detectStandard.}
}
\description{
UI for the default safetyGraphics shiny app
}
