% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyHistogram.R
\name{safetyHistogram}
\alias{safetyHistogram}
\title{Create a Safety Histogram widget}
\usage{
safetyHistogram(data, id_col = "USUBJID", value_col = "STRESN",
  measure_col = "TEST", normal_col_low = "STNRLO",
  normal_col_high = "STNRHI", unit_col = "STRESU", filters = NULL,
  details = NULL, start_value = NULL, missingValues = c("", "NA",
  "N/A"), debug_js = FALSE, settings = NULL)
}
\arguments{
\item{data}{A data frame containing the labs data. Data must be structured as one record per study participant per time point per lab measure.}

\item{id_col}{Unique subject identifier variable name. Default: \code{"USUBJID"}.}

\item{value_col}{Lab result variable name. Default: \code{"STRESN"}.}

\item{measure_col}{Lab measure variable name. Default: \code{"TEST"}.}

\item{normal_col_low}{Lower limit of normal variable name. Default: \code{"STNRLO"}.}

\item{normal_col_high}{Upper limit of normal variable name. Default: \code{"STNRHI"}.}

\item{unit_col}{Unit of measure variable name. Default is \code{"STRESU"}.}

\item{filters}{An optional list of specifications for filters.  Each filter is a nested, named list (containing the filter value column: "value_col" and associated label: "label") within the larger list. Default: \code{NULL}.}

\item{details}{An optional list of specifications for details listing.  Each column to be added to details listing is a nested, named list (containing the variable name: "value_col" and associated label: "label") within the larger list. Default: \code{NULL}.}

\item{start_value}{Value of variable defined in \code{measure_col} to be rendered in the histogram when the widget loads.}

\item{missingValues}{Vector of values defining a missing \code{value_col}. Default is \code{c('','NA','N/A')}.}

\item{debug_js}{print settings in javascript before rendering chart. Default: \code{FALSE}.}

\item{settings}{Optional list of settings arguments to be converted to JSON using \code{jsonlite::toJSON(settings, auto_unbox = TRUE, dataframe = "rows", null = "null")}.  If provided, all other function parameters are ignored. Default: \code{NULL}.}
}
\description{
This function creates a Safety Histogram using R htmlwidgets.
}
\examples{
\dontrun{

## Create Histogram figure customized to user data
safetyHistogram(data=adlbc, 
               id_col = "USUBJID",
               value_col = "AVAL", 
               measure_col = "PARAM", 
               normal_col_low = "A1LO", 
               normal_col_high = "A1HI", 
               unit_col = "PARAMCD")

## Create Histogram figure using a premade settings list
details_list <- list(
  list(value_col = "TRTP", label = "Treatment"),
  list(value_col = "SEX", label = "Sex"), 
  list(value_col = "AGEGR1", label = "Age group")
)


filters_list <- list(
  list(value_col = "TRTA", label = "Treatment"),
  list(value_col = "SEX", label = "Sex"), 
  list(value_col = "RACE", label = "RACE"),
  list(value_col = "AGEGR1", label = "Age group")
)

settingsl <- list(id_col = "USUBJID",
      value_col = "AVAL", 
      measure_col = "PARAM", 
      unit_col = "PARAMCD",
      normal_col_low = "A1LO", 
      normal_col_high = "A1HI", 
      details = details_list,
      filters = filters_list)
      
safetyHistogram(data=adlbc, settings = settingsl)

}

}
