% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyoutlierexplorer.R
\name{safetyoutlierexplorer}
\alias{safetyoutlierexplorer}
\title{Create a Safety Ourlier Explorer widget}
\usage{
safetyoutlierexplorer(data, debug_js = FALSE, settings = NULL)
}
\arguments{
\item{data}{A data frame containing the labs data. Data must be structured as one record per study participant per time point per lab measure.}

\item{debug_js}{print settings in javascript before rendering chart. Default: \code{FALSE}.}

\item{settings}{Optional list of settings arguments to be converted to JSON using \code{jsonlite::toJSON(settings, auto_unbox = TRUE, dataframe = "rows", null = "null")}. Default: \code{NULL}.}
}
\description{
This function creates an \href{https://github.com/rhoinc/safety-outlier-explorer/}{safety outlier explorer}. See the \href{https://github.com/rhoinc/safety-outlier-explorer/wiki/Configuration}{chart docuemtnation} for details regarding the settings object.
}
\examples{
\dontrun{

## Create Outlier Explorer figure using a premade settings list
details_list <- list(
  list(value_col = "TRTP", label = "Treatment"),
  list(value_col = "SEX", label = "Sex"),
  list(value_col = "AGEGR1", label = "Age group")
)


filters_list <- list(
  list(value_col = "TRTA", label = "Treatment"),
  list(value_col = "SEX", label = "Sex"),
  list(value_col = "RACE", label = "RACE"),
  list(value_col = "AGEGR1", label = "Age group")
)

settingsl <- list(id_col = "USUBJID",
      value_col = "AVAL",
      measure_col = "PARAM",
      unit_col = "PARAMCD",
      normal_col_low = "A1LO",
      normal_col_high = "A1HI",
      details = details_list,
      filters = filters_list)

safetyoutlierexplorer(data=adlbc, settings = settingsl)

}

}
