% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSettingsValue.R
\name{setSettingsValue}
\alias{setSettingsValue}
\title{Set the value for a given named parameter}
\usage{
setSettingsValue(key, value, settings)
}
\arguments{
\item{key}{a list (like those provided by \code{getSettingKeys()}) defining the position of parameter in the settings object.}

\item{value}{the value to set}

\item{settings}{The settings list used to generate a chart like \code{eDISH()}}
}
\value{
the updated settings object
}
\description{
Sets the value for a named parameter (\code{key}) to given \code{value} in a list (\code{settings})
}
\examples{
safetyGraphics:::setSettingValue(key=list("a","b"), value="notMyValue", settings=list(a=list(b="myValue"))) #returns list(a=list(b="notMyValue")))

testSettings<-generateSettings(standard="AdAM")
safetyGraphics:::setSettingValue(list("id_col"),"customID",testSettings) 
safetyGraphics:::setSettingValue(list("measure_values","ALP"),"Alanine Aminotransferase",testSettings)
safetyGraphics:::setSettingValue(list("myCustomSetting"),"customized",testSettings) #adds myCustomSetting to testSettings

}
\keyword{internal}
