% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settingsCharts.R
\name{settingsCharts}
\alias{settingsCharts}
\title{Server for settings tab showing details for the charts loaded in the app}
\usage{
settingsCharts(input, output, session, charts)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{charts}{list data frame summarizing the charts}
}
\description{
Server for settings tab showing details for the charts loaded in the app
}
