% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settingsCharts.R
\name{settingsChartsUI}
\alias{settingsChartsUI}
\title{UI for settings tab showing details for the charts loaded in the app}
\usage{
settingsChartsUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
UI for settings tab showing details for the charts loaded in the app
}
