% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settingsCode.R
\name{settingsCodeUI}
\alias{settingsCodeUI}
\title{UI for settings tab providing code to re-start the app with current data/settings}
\usage{
settingsCodeUI(id)
}
\arguments{
\item{id}{module ID}
}
\description{
UI for settings tab providing code to re-start the app with current data/settings
}
