% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settingsData.R
\name{settingsData}
\alias{settingsData}
\title{Settings Module - data details - server}
\usage{
settingsData(input, output, session, allData)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{allData}{tibble containing the data used in the app. Columns should be c('domain','name','standard','df')}
}
\value{
dataList - this may include files uploaded by the user.
}
\description{
server for the display of the loaded data
}
