% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settingsData.R
\name{settingsDataUI}
\alias{settingsDataUI}
\title{UI for settings tab showing current data}
\usage{
settingsDataUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
UI for settings tab showing current data
}
