% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settingsMapping.R
\name{settingsMapping}
\alias{settingsMapping}
\title{Server for settings tab showing current mapping}
\usage{
settingsMapping(input, output, session, metadata, mapping)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{metadata}{Data mapping metadata used for initial loading of app}

\item{mapping}{reactive data frame representing the current metadata mapping. columns = "domain", "text_id" and "current"}
}
\description{
Server for settings tab showing current mapping
}
