% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settingsMapping.R
\name{settingsMappingUI}
\alias{settingsMappingUI}
\title{Settings view of Metadata/Mapping - UI}
\usage{
settingsMappingUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
UI that displays the data mapping metadata.
}
