% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settingsMetadata.R
\docType{data}
\name{settingsMetadata}
\alias{settingsMetadata}
\title{settingsMetadata Dataset}
\format{A data frame}
\source{
Rho, Inc.
}
\usage{
settingsMetadata
}
\description{
A dataset containing things that Becca and Jeremy will better explain
}
\keyword{datasets}
