% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settingsTab.R
\name{settingsTab}
\alias{settingsTab}
\title{Server for the setting page}
\usage{
settingsTab(input, output, session, domains, metadata, mapping, charts)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{domains}{domains}

\item{metadata}{metadata}

\item{mapping}{mapping}

\item{charts}{charts}
}
\description{
Server for the setting page
}
