% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textKeysToList.R
\name{textKeysToList}
\alias{textKeysToList}
\title{Helper function to convert keys from text to nested lists}
\usage{
textKeysToList(textKeys)
}
\arguments{
\item{textKeys}{a list (or vector) of character keys using the "--" delimiter to indicate hierarchy}
}
\value{
A list of unnamed lists, with position in the nested list indicating hierarchy
}
\description{
Convert settings keys from text vectors (using the "--" delimiter) to a list of lists
}
\examples{
safetyGraphics:::textKeysToList("id_col") 
#list(list("id_col"))

#list(list("id_col"),list("measure_col","label"))
safetyGraphics:::textKeysToList(c("id_col","measure_col--label")) 

}
\keyword{internal}
