% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimSettings.R
\name{trimSettings}
\alias{trimSettings}
\title{Subset a settings object to thsoe relevant for a list of charts}
\usage{
trimSettings(settings, charts = NULL)
}
\arguments{
\item{settings}{The settings list to subset}

\item{charts}{The charts to subset by}
}
\value{
A list containing settings subsetted for the selected charts
}
\description{
This function returns a settings object
}
\details{
This function returns a settings object subsetted to the settings relevant to a vector of charts
}
\examples{
testSettings <- generateSettings(standard="adam")
trimSettings(settings=testSettings, charts = c("safetyhistogram","edish"))

}
