% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Server for the default safetyGraphics shiny app}
\usage{
app_server(input, output, session, meta, mapping, domainData, charts)
}
\arguments{
\item{input}{app input}

\item{output}{app output}

\item{session}{app session}

\item{meta}{data frame containing the metadata for use in the app. See the preloaded file (\code{?safetyGraphics::meta}) for more data specifications and details. Defaults to \code{safetyGraphics::meta}.}

\item{domainData}{named list of data.frames to be loaded in to the app.}

\item{chartsList}{list of charts to include in the app}
}
\description{
This function returns a server function suitable for use in shiny::runApp()
}
