% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_startup.R
\name{app_startup}
\alias{app_startup}
\title{Startup code for shiny app}
\usage{
app_startup(
  domainData = NULL,
  meta = NULL,
  charts = NULL,
  mapping = NULL,
  filterDomain = NULL,
  chartSettingsPaths = NULL
)
}
\arguments{
\item{domainData}{named list of domain level data}

\item{meta}{list of configuration metadata}

\item{charts}{list of charts}

\item{mapping}{initial data mapping}

\item{filterDomain}{domain used for the data/filter tab. Demographics ("\code{dm}") is used by default. Using a domain that is not one record per participant is not recommended.}

\item{chartSettingsPaths}{character vector with paths to chart setting yaml files}
}
\value{
List of elements for used to initialize the shiny app with the following parameters
\itemize{
\item{"meta"}{ List of configuration metadata }
\item{"charts"}{ List of charts }
\item{"domainData"}{ List of domain level data sets }
\item{"mapping"}{ Initial Data Mapping  }
\item{"standards"}{ List of domain level data standards }
}
}
\description{
Prepare inputs for safetyGraphics app - run before app is initialized. See ?safetyGraphicsApp for parameter definitions
}
