% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chartsNav.R
\name{chartsNav}
\alias{chartsNav}
\title{Add a navbar tab that initializes the Chart Module UI}
\usage{
chartsNav(chart, ns)
}
\arguments{
\item{chart}{chart metadata}

\item{ns}{namespace}
}
\description{
Add a navbar tab that initializes the Chart Module UI
}
