% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chartsRenderModule.R
\name{chartsRenderModule}
\alias{chartsRenderModule}
\title{Charts Module - render static chart server}
\usage{
chartsRenderModule(input, output, session, serverFunction, params)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{params}{parameters to be passed to the widget \link{REACTIVE}}

\item{chartFunction}{function to generate the chart.}
}
\description{
server for the display of the loaded data
}
