% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chartsRenderStatic.R
\name{chartsRenderStatic}
\alias{chartsRenderStatic}
\title{Charts Module - render static chart server}
\usage{
chartsRenderStatic(
  input,
  output,
  session,
  chartFunction,
  initFunction,
  data,
  mapping
)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{chartFunction}{function to generate the chart.}

\item{initFunction}{function called before the chart is generated. Should return a list of parameters that will be provided to chartFunction. returns list(data=data, settings=settings) by default.}

\item{data}{named list of data sets [reactive]}

\item{mapping}{data mapping [reactive]}
}
\description{
server for the display of the loaded data
}
