% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chartsRenderWidget.R
\name{chartsRenderWidget}
\alias{chartsRenderWidget}
\title{Charts Module - render static chart server}
\usage{
chartsRenderWidget(
  input,
  output,
  session,
  chart,
  package,
  params,
  settingsToJSON = TRUE
)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{chart}{chart name - must match the name of a widget in the specified pacakge}

\item{package}{package containing the widget. Note that package name is required for htmlwidgets.}

\item{params}{parameters to be passed to the widget [REACTIVE]}

\item{settingsToJSON}{convert param$settings to json? Default = TRUE}
}
\description{
server for the display of the loaded data
}
