% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chartsTab.R
\name{chartsTabUI}
\alias{chartsTabUI}
\title{UI for chart module, designed to be re-used for each chart generated.}
\usage{
chartsTabUI(id, chart)
}
\arguments{
\item{id}{module id}

\item{chart}{list containing chart specifications like those returned by \link{makeChartConfig}.}
}
\description{
UI for chart module, designed to be re-used for each chart generated.
}
