% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectStandard.R
\name{detectStandard}
\alias{detectStandard}
\title{Detect the data standard used for a data set}
\usage{
detectStandard(data, domain = "labs")
}
\arguments{
\item{data}{A data frame in which to detect the data standard}

\item{domain}{The data domain for the data set provided.  Default: \code{"labs"}.}
}
\value{
A list containing the matching \code{"standard"} ("ADaM", "SDTM" or "None") and a list of  \code{"details"} descripting of comparison with each standard considered. #' @examples 
detectStandard(adlbc)[["standard"]] #AdAM
detectStandard(iris)[[standard]] #none

\dontrun{
  detectStandard(adlbc,domain="AE") #throws error. AE domain not supported in this release. 
}
}
\description{
This function attempts to detect the data CDISC clinical standard used in a given R data frame.
}
\details{
This function compares the columns in the provided \code{"data"} with the required columns for a given data standard/domain combination. The function is designed to work with the SDTM and AdAM CDISC(<https://www.cdisc.org/>) standards for clinical trial data. Currently, only "labs" is the only domain supported.
}
