% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDISH.R
\name{eDISH}
\alias{eDISH}
\title{Create an eDISH widget}
\usage{
eDISH(data)
}
\arguments{
\item{A}{data frame containing the labs data. Data must be structured as one record per study participant per time point per lab measure.}
}
\description{
This function creates an interactive graphic for the Evaluation of Drug-Induced Serious Hepatotoxicity (eDISH)
}
