% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateStandard.R
\name{evaluateStandard}
\alias{evaluateStandard}
\title{Evaluate a data set against a data standard}
\usage{
evaluateStandard(data, standard, includeFields = TRUE, dmeta = NULL)
}
\arguments{
\item{data}{data.frame to evaluate}

\item{standard}{standard to evaluate}

\item{includeFields}{should field level data be evaluated?}

\item{dmeta}{domain level metadata, like that generated by \code(generateMetadata(domain="labs"))}
}
\value{
a list describing to what degree the data set matches the data standard. The "match" property describes compliance with the standard as "full", "partial" or "none". The "checks" property is a list of the data elements expected for the standard and whether they are "valid" in the given data set. "total_checks", "valid_checks" and "invalid_checks" provide counts of the specified checks. "match_percent" is calculated as valid_checks/total_checks.
}
\description{
Determines whether the required data elements in a data standard are found in a given data frame
}
\examples{
labs_meta <- getMetadata(domain="labs")
safetyGraphics:::evaluateStandard(data=labs, standard="adam", dmeta=labs_meta) # Match is TRUE
safetyGraphics:::evaluateStandard(data=labs, standard="sdtm", dmeta=labs_meta) # Match is FALSE

}
\keyword{internal}
