% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filterTab.R
\name{filterTab}
\alias{filterTab}
\title{filter module server}
\usage{
filterTab(
  input,
  output,
  session,
  domainData,
  filterDomain,
  current_mapping,
  tabID = "Filtering"
)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{domainData}{list of data files for each domain}

\item{filterDomain}{domain to use for filtering (typically "dm")}

\item{current_mapping}{current data mapping}

\item{tabID}{ID for the tab containing the filter UI (used for testing)}
}
\value{
filtered data set
}
\description{
server function that facilitates the data filtering with the datamods::filter_data_server module
}
