% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filterTab.R
\name{filterTab}
\alias{filterTab}
\title{Server for the filter module in datamods::filter_data_ui}
\usage{
filterTab(
  input,
  output,
  session,
  domainData,
  filterDomain,
  current_mapping,
  tabID = "Filtering",
  filterVars = NULL
)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{domainData}{list of data files for each domain}

\item{filterDomain}{domain to use for filtering (typically "dm")}

\item{current_mapping}{current data mapping}

\item{tabID}{ID for the tab containing the filter UI (used for testing)}

\item{filterVars}{Variables to use for filtering (used for testing)}
}
\value{
filtered data set
}
\description{
Server for the filter module in datamods::filter_data_ui
}
