% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filterTab.R
\name{filterTabUI}
\alias{filterTabUI}
\title{filterTabUI}
\usage{
filterTabUI(id, filterDomain = "dm")
}
\arguments{
\item{id}{module id}

\item{filterDomain}{data set for the domain}
}
\description{
UI that facilitates the filtering data with esquisse::filterDF_UI
}
