% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMappingList.R
\name{generateMappingList}
\alias{generateMappingList}
\title{Convert mapping data.frame to a list}
\usage{
generateMappingList(settingsDF, domain)
}
\arguments{
\item{domain}{mapping domain to return (returns all domains as a named list by default)}

\item{mappingDF}{data frame containing current mapping}
}
\description{
Convert mapping data.frame to a list
}
