% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateShell.R
\name{generateShell}
\alias{generateShell}
\title{Generate a default settings shell based on settings metadata}
\usage{
generateShell(charts = NULL, dmeta = NULL)
}
\arguments{
\item{charts}{The chart or chart(s) to include in the shell settings object}

\item{dmeta}{domain level metadata, like that generated by \code{generateMetadata(domain="labs")}}
}
\value{
A list containing a setting shell (all values = NA) for the selected chart(s)
}
\description{
This function returns a default settings object based on the chart(s) specified.
}
\details{
The function is designed to work with valid safetyGraphics charts.
}
\examples{

safetyGraphics:::generateShell(charts = "hepexplorer", dmeta=getMetadata(domain="labs"))

}
\keyword{internal}
