% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRequiredSettings.R
\name{getRequiredSettings}
\alias{getRequiredSettings}
\title{Get a list of required settings}
\usage{
getRequiredSettings(charts = NULL, dmeta = null)
}
\arguments{
\item{charts}{The chart for which required settings should be returned ("eDish" only for now) .  Default: \code{NULL} (uses all available charts).}

\item{dmeta}{domain level metadata, like that generated by \code{generateMetadata(domain="labs")}}
}
\value{
List of lists specifying the position of matching named elements in the format \code{list("filters",2,"value_col")}, which would correspond to \code{settings[["filters"]][[2]][["value_col"]]}.
}
\description{
Get a list of required settings for a given chart
}
\examples{
labs_meta<-getMetadata(domain="labs")
safetyGraphics:::getRequiredSettings(charts=c("edish","safetyHistogram"),dmeta=labs_meta)

}
