% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRequiredSettings.R
\name{getRequiredSettings}
\alias{getRequiredSettings}
\title{Get a list of required settings}
\usage{
getRequiredSettings(charts = NULL, metadata = safetyGraphics::settingsMetadata)
}
\arguments{
\item{charts}{The chart for which required settings should be returned ("eDish" only for now) .  Default: \code{NULL} (uses all available charts).}

\item{metadata}{The metadata file to be used (primarily used for testing)}
}
\value{
List of lists specifying the position of matching named elements in the format \code{list("filters",2,"value_col")}, which would correspond to \code{settings[["filters"]][[2]][["value_col"]]}.
}
\description{
Get a list of required settings for a given chart
}
\examples{
safetyGraphics:::getRequiredSettings(charts=c("edish","safetyHistogram"))

}
