% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_homeTab.R
\name{homeTab}
\alias{homeTab}
\title{Server for the filter module in datamods::filter_data_ui}
\usage{
homeTab(input, output, session)
}
\arguments{
\item{input}{mod input}

\item{output}{mod output}

\item{session}{mod session}
}
\description{
Server for the filter module in datamods::filter_data_ui
}
