% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loadCharts.R
\name{loadCharts}
\alias{loadCharts}
\title{loadCharts}
\usage{
loadCharts(input, output, session, charts = makeChartConfig())
}
\arguments{
\item{domains}{List of data domains to be loaded}
}
\description{
server that facilitates selection of charts for safetyGraphicsApp
}
