% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loadCharts.R
\name{loadCharts}
\alias{loadCharts}
\title{loadCharts}
\usage{
loadCharts(input, output, session, charts = makeChartConfig())
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{charts}{Initial list of charts}
}
\description{
server that facilitates selection of charts for safetyGraphicsApp
}
