% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loadCharts.R
\name{loadChartsUI}
\alias{loadChartsUI}
\title{loadChartsUI}
\usage{
loadChartsUI(id, charts = makeChartConfig())
}
\arguments{
\item{id}{module id}

\item{charts}{list of charts}
}
\description{
UI that helps users pick charts
}
