% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loadData.R
\name{loadData}
\alias{loadData}
\title{loadDataServer}
\usage{
loadData(input, output, session, domain)
}
\arguments{
\item{domain}{List of data domains to be loaded}
}
\description{
UI that facilitates the mapping of a column data domain
}
