% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loadData.R
\name{loadDataUI}
\alias{loadDataUI}
\title{UI for the data loading module used in safetyGraphicsInit()}
\usage{
loadDataUI(id, domain = NULL)
}
\arguments{
\item{id}{module id}

\item{domain}{character vector with domains to be loaded}
}
\description{
UI for the data loading module used in safetyGraphicsInit()
}
