% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeChartConfig.R
\name{makeChartConfig}
\alias{makeChartConfig}
\title{Make Chart Config}
\usage{
makeChartConfig(
  dirs,
  packages = "safetyCharts",
  packageLocation = "config",
  sourceFiles = TRUE
)
}
\arguments{
\item{dirs}{path to one or more directories containing yaml config files (relative to working directory)}

\item{packages}{installed packages names containing yaml config files in the /inst/{packageLocation} folder}

\item{packageLocation}{inst folder where yaml config files are located in packages}

\item{sourceFiles}{boolean indicating whether to source all R files found in dirs.}
}
\value{
returns a named list of charts derived from YAML files. Each element of the list contains information about a single chart, and has the following parameters:
\itemize{
\item{"name"}{ Name of the chart. Also the name of the element in the list - e.g. charts$aeExplorer$name is "aeExplorer"}
\item{"label"}{ short description of the chart }
\item{"type"}{ type of chart; options are: 'htmlwidget', 'module', 'plot', 'table', 'html' or 'plotly'.}
\item{"domain"}{ data domain. Should correspond to a domain in \code{meta} or be set to "multiple" }
\item{"package"}{ primary package (if any). Other packages can be loaded directly in workflow functions. }
\item{"path"}{ Path to YAML file}
\item{"workflow"}{ List of functions names used to render chart. See vignette for details. }
\item{"functions"}{ List of functions for use in chart renderering. These functions must be located in the global environment or \code{package} field of the YAML config. Function names must include either the \code{name} or \code{workflow} fields of the YAML config. }
}
}
\description{
Converts YAML chart configuration files to an R list and binds workflow functions. See the vignette about creating custom charts for more details.
}
