% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeChartExport.R
\name{makeChartExport}
\alias{makeChartExport}
\title{Make Chart Config Functions}
\usage{
makeChartExport(chart, mapping)
}
\arguments{
\item{chart}{chart object like the one generated by makeChartConfig().}
}
\value{
returns the chart object with an export script added.
}
\description{
Binds needed functions to a chart object based on chart type.
}
