% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeChartSummary.R
\name{makeChartSummary}
\alias{makeChartSummary}
\title{html chart summary}
\usage{
makeChartSummary(
  chart,
  status = NULL,
  showLinks = TRUE,
  class = "chart-header"
)
}
\arguments{
\item{chart}{list containing chart specifications}

\item{status}{(optional) chart status from \code{getChartStatus}. Default is NULL.}

\item{showLinks}{boolean indicating whether to include links}

\item{class}{character to include as class}
}
\description{
makes a nicely formatted html summary for a chart object
}
