% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMeta.R
\name{makeMeta}
\alias{makeMeta}
\title{Create a metadata object table for a set of charts}
\usage{
makeMeta(charts)
}
\arguments{
\item{charts}{list of safetyGraphics chart objects for which to create metadata}
}
\value{
tibble of metadata with the following columns:
\describe{
\item{domain}{Data domain}
\item{text_key}{Text key indicating the setting name. \code{'--'} delimiter indicates a field level data mapping}
\item{col_key}{Key for the column mapping}
\item{field_key}{Key for the field mapping (if any)}
\item{type}{type of mapping - "field" or "column"}
\item{label}{Label}
\item{description}{Description}
\item{multiple}{Mapping supports multiple columns/fields }
\item{standard_adam}{Default values for the ADaM data standard}
\item{standard_sdtm}{Default values for the SDTM data standard}
}
}
\description{
Generates metadata object for a list charts. makeMeta() looks in chart$package namespace for files called meta_{chart$name} and meta_{chart$domain} for all charts, and then stacks all files. If duplicate metadata rows (domain + text_key) are found an error is thrown.
}
