% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mappingColumn.R
\name{mappingColumn}
\alias{mappingColumn}
\title{mappingColumn}
\usage{
mappingColumn(input, output, session, meta, data)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{meta}{metadata data frame for the object}

\item{data}{current data file for the domain}
}
\value{
A reactive data.frame providing the current value for text_key associated with the selected column
}
\description{
server function that facilitates the mapping of a single data element (column of field) with a simple select UI
}
