% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mappingDomain.R
\name{mappingDomain}
\alias{mappingDomain}
\title{mappingSelect}
\usage{
mappingDomain(input, output, session, meta, data)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\value{
A reactive data frame containing the mapping for the domain
}
\description{
server function that facilitates the mapping of a single data element (column of field) with a simple select UI
}
