% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mappingTab.R
\name{mappingTab}
\alias{mappingTab}
\title{Server for mapping tab covering of all data domains}
\usage{
mappingTab(input, output, session, meta, domainData)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{meta}{metadata for all domains}

\item{domainData}{clinical data for all domains}
}
\value{
list of mappings for all domains
}
\description{
Server for mapping tab covering of all data domains
}
