% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mappingTab.R
\name{mappingTabUI}
\alias{mappingTabUI}
\title{mappingTabUI}
\usage{
mappingTabUI(id, meta, domainData, mappings = NULL, standards = NULL)
}
\arguments{
\item{id}{module id}

\item{meta}{metadata for all domains}

\item{domainData}{list of data files for each domain}

\item{mappings}{optional data frame containing stacked mappings for all domains}

\item{standards}{optional list of data standards like the ones generated by \code{detectStandard()}}
}
\description{
UI that facilitates the mapping of all data domain
}
