% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSettings.R
\name{removeSettings}
\alias{removeSettings}
\title{Remove a setting from the safetyGraphics shiny app}
\usage{
removeSettings(text_keys, settingsLocation = getwd())
}
\arguments{
\item{text_keys}{Text keys indicating the setting names to be removed.}

\item{settingsLocation}{path where the custom settings will be loaded/saved. If metadata is not found in that location, it will be read from the package (e.g. safetyGraphics::settingsMetadata), and then written to the specified location once the setting has been removed.}
}
\description{
This function updates settings objects to remove a setting parameter from the safetyGraphics shiny app
}
\details{
This function makes it easy for remove a setting from the safetyGraphics shiny app by making updates to the underlying metadata used by the package.
}
