% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyGraphicsApp.R
\name{safetyGraphicsApp}
\alias{safetyGraphicsApp}
\title{Run the interactive safety graphics builder}
\usage{
safetyGraphicsApp(charts = NULL, maxFileSize = NULL,
  settingsLocation = NULL)
}
\arguments{
\item{charts}{Character vector of charts to include}

\item{maxFileSize}{maximum file size in MB allowed for file upload}

\item{settingsLocation}{folder location of user-defined settings metadata}
}
\description{
Run the interactive safety graphics builder
}
