% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyGraphicsApp.R
\name{safetyGraphicsApp}
\alias{safetyGraphicsApp}
\title{Run the interactive safety graphics builder}
\usage{
safetyGraphicsApp(
  maxFileSize = NULL,
  meta = safetyGraphics::meta,
  domainData = list(labs = safetyGraphics::labs, aes = safetyGraphics::aes),
  mapping = NULL
)
}
\arguments{
\item{maxFileSize}{maximum file size in MB allowed for file upload}

\item{meta}{data frame containing the metadata for use in the app. See the preloaded file (\code{?safetyGraphics::meta}) for more data specifications and details. Defaults to \code{safetyGraphics::meta}.}

\item{domainData}{mamed list of data.frames to be loaded in to the app.}

\item{mapping}{data.frame specifying the initial values for each data mapping. If no mapping is provided, the app will attempt to generate one via \code{detectStandard()}}
}
\description{
Run the interactive safety graphics builder
}
