% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyGraphicsApp.R
\name{safetyGraphicsApp}
\alias{safetyGraphicsApp}
\title{Run the interactive safety graphics builder}
\usage{
safetyGraphicsApp(
  maxFileSize = NULL,
  meta = safetyGraphics::meta,
  domainData = list(labs = safetyGraphics::labs, aes = safetyGraphics::aes),
  mapping = NULL
)
}
\arguments{
\item{maxFileSize}{maximum file size in MB allowed for file upload}

\item{meta}{data frame containing the metadata for use in the app. See the preloaded file (\code{?metadata}) for more data specifications and details. Defaults to \code{safetygraphics::metadata}.}

\item{domainData}{list of data.frames to be loaded in to the app -}
}
\description{
Run the interactive safety graphics builder
}
