% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyGraphicsApp.R
\name{safetyGraphicsApp}
\alias{safetyGraphicsApp}
\title{Run the interactive safety graphics builder}
\usage{
safetyGraphicsApp(maxFileSize = NULL)
}
\arguments{
\item{maxFileSize}{maximum file size in MB allowed for file upload.}
}
\description{
Run the interactive safety graphics builder
}
