% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyGraphicsInit.R
\name{safetyGraphicsInit}
\alias{safetyGraphicsInit}
\title{App to select charts, load data and then initialize the core safetyGraphics app}
\usage{
safetyGraphicsInit(
  charts = makeChartConfig(),
  delayTime = 1000,
  maxFileSize = NULL
)
}
\arguments{
\item{charts}{chart object}

\item{delayTime}{time (in ms) between drawing app UI and starting server. Default set to 1000 (1 second), but could need to be higher on slow machine.}

\item{maxFileSize}{maximum file size in MB allowed for file upload}
}
\description{
App to select charts, load data and then initialize the core safetyGraphics app
}
