% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyGraphicsInit.R
\name{safetyGraphicsInit}
\alias{safetyGraphicsInit}
\title{Run the interactive safety graphics app}
\usage{
safetyGraphicsInit(charts = makeChartConfig(), delayTime = 1000)
}
\arguments{
\item{charts}{chart object}
}
\description{
Run the interactive safety graphics app
}
