% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_safetyGraphicsServer.R
\name{safetyGraphicsServer}
\alias{safetyGraphicsServer}
\title{Server for core safetyGraphics app including Home, Mapping, Filter, Charts and Settings modules.}
\usage{
safetyGraphicsServer(
  input,
  output,
  session,
  meta,
  mapping,
  domainData,
  charts,
  filterDomain,
  config
)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{meta}{data frame containing the metadata for use in the app.}

\item{mapping}{current mapping}

\item{domainData}{named list of data.frames to be loaded in to the app.}

\item{charts}{list of charts to include in the app}

\item{filterDomain}{domain used for the data/filter tab. Demographics ("\code{dm}") is used by default. Using a domain that is not one record per participant is not recommended.}
}
\description{
This function returns a server function suitable for use in shiny::runApp()
}
