% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settingsCharts.R
\name{settingsCharts}
\alias{settingsCharts}
\title{Settings Module - charts details - server}
\usage{
settingsCharts(input, output, session, charts)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{charts}{list data frame summarizing the charts}
}
\description{
server for the display of the charts
}
