% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settingsCharts.R
\name{settingsChartsUI}
\alias{settingsChartsUI}
\title{Settings Module - chart details}
\usage{
settingsChartsUI(id)
}
\description{
Settings Module - sub-module showing details for the charts loaded in the app - UI
}
