% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settingsCode.R
\name{settingsCode}
\alias{settingsCode}
\title{Server for settings tab providing code to re-start the app with current data/settings}
\usage{
settingsCode(input, output, session, mapping, charts, domainData)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{mapping}{mapping}

\item{charts}{charts}

\item{domainData}{data list}
}
\description{
Server for settings tab providing code to re-start the app with current data/settings
}
