% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settingsCode.R
\name{settingsCode}
\alias{settingsCode}
\title{Server for the setting code page}
\usage{
settingsCode(id, input, output, session, mapping)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{mapping}{mapping}

\item{domains}{domains}

\item{metadata}{metadata}

\item{charts}{charts}
}
\description{
Server for the setting code page
}
