% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settingsData.R
\name{settingsData}
\alias{settingsData}
\title{Server for settings tab showing current data}
\usage{
settingsData(input, output, session, domains)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{domains}{named list of the data.frames for each domain}
}
\description{
Server for settings tab showing current data
}
