% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settingsData.R
\name{settingsDataUI}
\alias{settingsDataUI}
\title{Settings Module - data details}
\usage{
settingsDataUI(id)
}
\description{
Settings Module - sub-module showing details for the data loaded in the app - UI
}
