% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settingsMapping.R
\name{settingsMapping}
\alias{settingsMapping}
\title{Settings view of Metadata/Mapping - server}
\usage{
settingsMapping(input, output, session, metaIn, mapping)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{metaIn}{Data mapping metadata used for initial loading of app}

\item{mapping}{reactive data frame representing the current metadata mapping. columns = "domain", "text_id" and "current"}
}
\description{
server for the display of the data mapping metadata.
}
