% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settingsMapping.R
\name{settingsMappingUI}
\alias{settingsMappingUI}
\title{UI for settings tab showing current mapping}
\usage{
settingsMappingUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
UI for settings tab showing current mapping
}
