% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimData.R
\name{trimData}
\alias{trimData}
\title{Removes unnecessary rows and columns}
\usage{
trimData(data, settings, charts = NULL)
}
\arguments{
\item{data}{a data frame to trim}

\item{settings}{the settings list used to determine which rows and columns to drop}

\item{charts}{the charts being created  Default: \code{NULL} (include data needed for all available charts).}
}
\value{
A dataframe with unnecessary columns and rows removed
}
\description{
Removes unnecessary rows and columns from data based on current settings
}
\examples{
testSettings <- generateSettings(standard="adam")
trimmed<-safetyGraphics:::trimData(data=adlbc, settings=testSettings)

}
\keyword{internal}
